Update comment for namespace changes
I checked with Colin, apparently the //path:module syntax used to be
based on the path to the module, not the path to the namespace. I agree
that's better as it matches bazel and would help us phase out the
concept of namespaces, so update the comment to only suggest that option
and not the option of making the path the path of the namespace.

Bug: 432305765
Test: Presubmits
Change-Id: I9ffbca2fcccbbda842c60e9c0e2f4bc778e0257b